refactor: extract population subcomponents
