+++ src/components/SettlerFilters.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { Checkbox } from './ui/checkbox';
+
+export default function SettlerFilters({
+  onlyLiving,
+  setOnlyLiving,
+  unassignedOnly,
+  setUnassignedOnly,
+}) {
+  return (
+    <div className="flex flex-wrap items-center gap-4 text-sm">
+      <div className="flex items-center gap-2">
+        <Checkbox
+          id="onlyLiving"
+          checked={onlyLiving}
+          onCheckedChange={(v) => setOnlyLiving(v === true)}
+        />
+        <label htmlFor="onlyLiving">Only living</label>
+      </div>
+      <div className="flex items-center gap-2">
+        <Checkbox
+          id="unassignedOnly"
+          checked={unassignedOnly}
+          onCheckedChange={(v) => setUnassignedOnly(v === true)}
+        />
+        <label htmlFor="unassignedOnly">Unassigned only</label>
+      </div>
+    </div>
+  );
+}
